refactor: ナビゲーターの細分化
